refactors styling
