+++ frontend/src/Components/Pagination/components/Pagination/Pagination.module.scss
@@ -1,0 +1,4 @@
+.nav {
+  display: flex;
+  align-items: center;
+}
