+++ frontend/src/Components/Pagination/components/PaginationContent/PaginationContent.module.scss
@@ -1,0 +1,9 @@
+.list {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 0.5rem;
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
