+++ frontend/src/Components/Pagination/components/PaginationContent/PaginationContent.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-
+import styles from './PaginationContent.module.scss';
 export const PaginationContent = React.forwardRef<HTMLUListElement, React.ComponentProps<'ul'>>(
-  ({ className, ...props }, ref) => <ul ref={ref} className={className} {...props} />,
+  ({ className, ...props }, ref) => <ul ref={ref} className={`${styles.list} ${className}`} {...props} />,
 );
 PaginationContent.displayName = 'PaginationContent';
