+++ frontend/src/Components/Pagination/components/PaginationEllipsis/PaginationEllipsis.module.scss
@@ -1,0 +1,8 @@
+.ellipsis {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  min-width: 2rem;
+  letter-spacing: 2px;
+  user-select: none;
+}
