+++ frontend/src/Components/Pagination/components/PaginationEllipsis/PaginationEllipsis.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-
+import styles from './PaginationEllipsis.module.scss';
 export const PaginationEllipsis = React.forwardRef<HTMLSpanElement, React.ComponentProps<'span'>>(
   ({ className, ...props }, ref) => (
-    <span ref={ref} className={className} {...props}>
+    <span ref={ref} className={`${styles.ellipsis} ${className}`} {...props}>
       . . .
     </span>
   ),
