+++ frontend/src/Components/Pagination/components/PaginationItem/PaginationItem.module.scss
@@ -1,0 +1,4 @@
+.item {
+  display: flex;
+  align-items: center;
+}
