+++ frontend/src/Components/Pagination/components/index.ts
@@ -1,5 +1,5 @@
-export { Pagination } from './Pagination';
-export { PaginationContent } from './PaginationContent';
+export { Pagination } from './Pagination/Pagination';
+export { PaginationContent } from './PaginationContent/PaginationContent';
 export { PaginationControll } from './PaginationControll';
-export { PaginationEllipsis } from './PaginationEllipsis';
-export { PaginationItem } from './PaginationItem';
+export { PaginationEllipsis } from './PaginationEllipsis/PaginationEllipsis';
+export { PaginationItem } from './PaginationItem/PaginationItem';
